separate message macros from util macros
