+++ stratumv2/src/lib.rs
@@ -8,6 +8,8 @@ use std::io;
 
 #[macro_use]
 mod internal_macros;
+#[macro_use]
+mod internal_message_macros;
 
 /// Errors returned in the library.
 mod error;
